/Assignment1
Your client “Jalopies Are Us" is a small used car lot that is looking to expand their business. They have asked you to develop a system that tracks their auto sales. Your task is to start by developing a Vehicle class that will hold all data as it pertains to any vehicle they have on their lot. The second and third stage of the project will see additional features added.